Break Dashboard down into 3 separate modules (#63)
- `data` holds feature-specific data storage. In this case, it contains
the dashboard configuration repository.
- `logic` holds business logic and related classes. This is effectively
a bunch of use cases.
- `ui` holds the feature UI and navigation. This is the "top" of the
feature, which will be included in `app`.
