+++ app/src/main/java/com/boswelja/truemanager/MainApplication.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import com.boswelja.truemanager.auth.AuthModule
 import com.boswelja.truemanager.core.api.v2.ApiV2Module
 import com.boswelja.truemanager.apps.AppsModule
-import com.boswelja.truemanager.dashboard.DashboardModule
+import com.boswelja.truemanager.dashboard.ui.DashboardModule
 import com.boswelja.truemanager.reporting.ReportingModule
 import com.boswelja.truemanager.storage.StorageModule
 import org.koin.android.ext.koin.androidContext
