--- features/dashboard/build.gradle.kts
@@ -1,77 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.ksp)
-
-    alias(libs.plugins.detekt)
-}
-
-android {
-    namespace = "com.boswelja.truemanager.dashboard"
-    compileSdk = 33
-
-    defaultConfig {
-        minSdk = 28
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-
-    buildFeatures.compose = true
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.compose.compiler.get()
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    lint {
-        sarifReport = true
-        htmlReport = false
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-}
-
-kotlin {
-    jvmToolchain(17)
-}
-
-ksp {
-    arg("room.schemaLocation", "$projectDir/schemas")
-    arg("room.incremental", "true")
-    arg("room.generateKotlin", "true")
-}
-
-detekt {
-    buildUponDefaultConfig = true
-    config.setFrom("$rootDir/config/detekt.yml")
-    basePath = rootDir.absolutePath
-}
-
-dependencies {
-    implementation(projects.core.api)
-    implementation(projects.core.capacity)
-
-    implementation(libs.androidx.navigation)
-    implementation(libs.bundles.compose)
-    debugImplementation(libs.bundles.compose.tooling)
-
-    implementation(libs.vico)
-
-    implementation(libs.koin.android)
-    implementation(libs.koin.android.compose)
-
-    implementation(libs.room.runtime)
-    ksp(libs.room.compiler)
-}
