+++ features/dashboard/data/schemas/com.boswelja.truemanager.data.configuration.database.DashboardEntryDatabase/1.json
