+++ features/dashboard/data/src/main/AndroidManifest.xml
