+++ features/dashboard/logic/build.gradle.kts
@@ -1,0 +1,60 @@
+plugins {
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.kotlin.android)
+
+    alias(libs.plugins.detekt)
+}
+
+android {
+    namespace = "com.boswelja.truemanager.dashboard.logic"
+    compileSdk = 33
+
+    defaultConfig {
+        minSdk = 28
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+        }
+    }
+
+    packaging {
+        resources {
+            excludes += "/META-INF/{AL2.0,LGPL2.1}"
+        }
+    }
+    lint {
+        sarifReport = true
+        htmlReport = false
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+}
+
+kotlin {
+    jvmToolchain(17)
+}
+
+detekt {
+    buildUponDefaultConfig = true
+    config.setFrom("$rootDir/config/detekt.yml")
+    basePath = rootDir.absolutePath
+}
+
+dependencies {
+    api(projects.core.capacity)
+    api(libs.kotlinx.coroutines)
+    api(libs.kotlinx.datetime)
+
+    implementation(projects.core.api)
+    implementation(projects.features.dashboard.data)
+
+    implementation(libs.koin.core)
+}
