+++ features/dashboard/logic/src/main/kotlin/com/boswelja/truemanager/dashboard/logic/ExtractSystemInformationData.kt
@@ -1,0 +1,27 @@
+package com.boswelja.truemanager.dashboard.logic
+
+import com.boswelja.truemanager.core.api.v2.system.SystemInfo
+import kotlinx.datetime.Instant
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.toLocalDateTime
+
+/**
+ * Takes data received from the server and maps it to a [DashboardData.SystemInformationData]. See
+ * [invoke] for details.
+ */
+class ExtractSystemInformationData {
+
+    /**
+     * Produces a [DashboardData.SystemInformationData] from information retrieved from the server.
+     *
+     * @param systemInfo The server system information.
+     */
+    operator fun invoke(systemInfo: SystemInfo): DashboardData.SystemInformationData {
+        return DashboardData.SystemInformationData(
+            version = systemInfo.version,
+            hostname = systemInfo.hostName,
+            lastBootTime = Instant.fromEpochMilliseconds(systemInfo.bootTime)
+                .toLocalDateTime(TimeZone.currentSystemDefault())
+        )
+    }
+}
